+++ docs/src/TextArea.doc.js
@@ -56,6 +56,10 @@ card(
         type: '({ event: SyntheticFocusEvent<>, value: string }) => void',
       },
       {
+        name: 'onKeyDown',
+        type: '({ event: SyntheticKeyboardEvent<>, value: string }) => void',
+      },
+      {
         name: 'placeholder',
         type: 'string',
       },