+++ packages/gestalt/src/TextField.test.js
@@ -6,6 +6,10 @@ import TextField from './TextField.js';
 import Flyout from './Flyout.js';
 
 describe('TextField', () => {
+  beforeAll(() => {
+    // Mock this out for the instanceof checks in event handlers
+    global.HTMLInputElement = Object;
+  });
   it('Renders a Flyout if an error message is passed in', () => {
     const wrapper = shallow(
       <TextField errorMessage="test" id="test" onChange={jest.fn()} />
@@ -113,5 +117,51 @@ describe('TextField', () => {
       />
     ).html();
     expect(tree).toMatchSnapshot();
+  });
+
+  it('handles blur events', () => {
+    const mockBlur = jest.fn();
+    const tree = shallow(
+      <TextField id="test" onBlur={mockBlur} onChange={jest.fn()} />
+    );
+    tree.find('input').simulate('blur', { target: { value: 'fake value' } });
+    expect(mockBlur).toHaveBeenCalledWith({
+      event: { target: { value: 'fake value' } },
+      value: 'fake value',
+    });
+  });
+
+  it('handles change events', () => {
+    const mockChange = jest.fn();
+    const tree = shallow(<TextField id="test" onChange={mockChange} />);
+    tree.find('input').simulate('change', { target: { value: 'fake value' } });
+    expect(mockChange).toHaveBeenCalledWith({
+      event: { target: { value: 'fake value' } },
+      value: 'fake value',
+    });
+  });
+
+  it('handles focus events', () => {
+    const mockFocus = jest.fn();
+    const tree = shallow(
+      <TextField id="test" onChange={jest.fn()} onFocus={mockFocus} />
+    );
+    tree.find('input').simulate('focus', { target: { value: 'fake value' } });
+    expect(mockFocus).toHaveBeenCalledWith({
+      event: { target: { value: 'fake value' } },
+      value: 'fake value',
+    });
+  });
+
+  it('handles key down events', () => {
+    const mockKeyDown = jest.fn();
+    const tree = shallow(
+      <TextField id="test" onChange={jest.fn()} onKeyDown={mockKeyDown} />
+    );
+    tree.find('input').simulate('keyDown', { target: { value: 'fake value' } });
+    expect(mockKeyDown).toHaveBeenCalledWith({
+      event: { target: { value: 'fake value' } },
+      value: 'fake value',
+    });
   });
 });